fix: Changed inventory_quantity to IntergerField

+++ products/models.py
@@ -7,5 +7,5 @@ class Product(models.Model):
     title = models.CharField(max_length=255)
     description = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=8, decimal_places=2)
-    inventory_quantity = models.DecimalField()
+    inventory_quantity = models.IntegerField()
     web_pic = models.URLField(max_length=200)
